Reconfigurada la transición a modo oscuro desde Django
