+++ MelaoApp/templates/MelaoApp/base.html
@@ -13,7 +13,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Plus+Jakarta+Sans:ital,wght@0,200..800;1,200..800&display=swap" rel="stylesheet">
 </head>
-<body>
+<body {% if request.COOKIES.theme == 'dark' %}class="dark-mode"{% endif %}>
     {% block custom_header %}
     <header class="mobile-navbar">
         <a href="{% url 'MelaoApp:home' %}">
