+++ static/MelaoApp/css/chatList.css
@@ -46,3 +46,28 @@ main {
         background-color: #FCFAF7;
     }
 }
+
+body.dark-mode {
+    background-color: #211C12; /* Color de fondo oscuro */
+    color: #f0f0f0; /* Color de texto claro */
+}
+body.dark-mode .NavBar {
+    background-color: #211C12; /* Navbar oscuro */
+}
+body.dark-mode .Menu {
+    background-color: #473D24; /* Navbar oscuro */
+}
+/* Agrega reglas similares para otros elementos como h2, p, etc. */
+body.dark-mode h2 {
+    color: #f0f0f0;
+}
+body.dark-mode .person-name{
+    color:#f0f0f0;
+}
+body.dark-mode .SearchTabDiv{
+    background-color: #473D24;
+}
+body.dark-mode .SearchTabDiv input{
+    background-color: #473D24;
+    color: #f0f0f0;
+}
