+++ static/MelaoApp/css/modifyProfile.css
@@ -41,3 +41,21 @@
         background-color: #FCFAF7;
     }
 }
+
+body.dark-mode {
+    background-color: #211C12; 
+    color: #f0f0f0;
+}
+body.dark-mode .NavBar {
+    background-color: #211C12;
+}
+body.dark-mode .Menu {
+    background-color: #473D24; 
+}
+
+body.dark-mode h2 {
+    color: #f0f0f0;
+}
+body.dark-mode a, body.dark-mode a:visited {
+  color: #f0f0f0;
+}
