+++ static/MelaoApp/css/newPost.css
@@ -229,3 +229,25 @@ textarea:focus {
         background-color: #F2F2F2;
     }
 }
+
+body.dark-mode {
+    background-color: #211C12; 
+    color: #f0f0f0; 
+}
+body.dark-mode .NavBar {
+    background-color: #211C12; 
+}
+body.dark-mode .Menu {
+    background-color: #473D24;
+}
+
+body.dark-mode h2 {
+    color: #f0f0f0;
+}
+body.dark-mode .PostDiv {
+    background-color: #473D24;
+}
+body.dark-mode #PostTextArea  {
+    background-color: #473D24;
+    color: #f0f0f0;
+}
