+++ static/MelaoApp/css/signUpView.css
@@ -5,3 +5,24 @@
         width: 100%
     }
 }
+
+body.dark-mode {
+    background-color: #211C12; 
+    color: #f0f0f0; 
+}
+body.dark-mode .NavBar, .LoginPannel{
+    background-color: #211C12; 
+}
+
+body.dark-mode .Menu {
+    background-color: #473D24;
+}
+
+body.dark-mode h2 {
+    color: #f0f0f0;
+}
+
+body.dark-mode .UserDataInput{
+    background-color: #473D24;
+    color: #f0f0f0;
+}
