+++ static/MelaoApp/css/viewNotifications.css
@@ -54,3 +54,28 @@ main {
         background-color: #FCFAF7;
     }
 }
+
+body.dark-mode {
+    background-color: #211C12; 
+    color: #f0f0f0; 
+}
+body.dark-mode .NavBar {
+    background-color: #211C12; 
+}
+body.dark-mode .Menu {
+    background-color: #473D24;
+}
+
+body.dark-mode h2 {
+    color: #f0f0f0;
+}
+body.dark-mode .PersonName{
+    color:#f0f0f0;
+}
+body.dark-mode .SearchTabDiv{
+    background-color: #473D24;
+}
+body.dark-mode .SearchTabDiv input{
+    background-color: #473D24;
+    color: #f0f0f0;
+}
