+++ static/MelaoApp/css/welcome.css
@@ -108,3 +108,19 @@ footer p {
         margin-bottom: 12px;
     }
 }
+
+body.dark-mode {
+    background-color: #211C12; 
+    color: #f0f0f0; 
+}
+body.dark-mode .NavBar, .LoginPannel{
+    background-color: #211C12; 
+}
+
+body.dark-mode .Menu {
+    background-color: #473D24;
+}
+
+body.dark-mode h2 {
+    color: #f0f0f0;
+}
