+++ static/MelaoApp/js/chatView.js
@@ -1,16 +1,11 @@
 function setTheme(){
-    const body = document.body;
     const BackArrow = document.getElementById('Arrow-button');
 
     const applyTheme = (theme) => {
         if (theme === 'dark') {
-            body.classList.add('dark-mode');
-            localStorage.setItem('theme', 'dark');
             BackArrow.querySelector('img').src = '/MelaoApp/resources/DarkMode/BackArrow.png';
             
         } else {
-            body.classList.remove('dark-mode');
-            localStorage.setItem('theme', 'light');
             BackArrow.querySelector('img').src = '/MelaoApp/resources/BackArrow.png';
         }
     };
