+++ static/MelaoApp/js/modifyProfile.js
@@ -1,12 +1,9 @@
 function setTheme(){
-    const body = document.body;
     const BackArrow = document.getElementById('Arrow-button');
     const RightArrow = document.getElementsByClassName('option-container');
 
     const applyTheme = (theme) => {
         if (theme === 'dark') {
-            body.classList.add('dark-mode');
-            localStorage.setItem('theme', 'dark');
             BackArrow.querySelector('img').src = '/MelaoApp/resources/DarkMode/BackArrow.png';
             Array.from(RightArrow).forEach(el => {
                 const img = el.querySelector('img');
@@ -14,8 +11,6 @@ function setTheme(){
             });
             
         } else {
-            body.classList.remove('dark-mode');
-            localStorage.setItem('theme', 'light');
             BackArrow.querySelector('img').src = '/MelaoApp/resources/BackArrow.png';
             Array.from(RightArrow).forEach(el => {
                 const img = el.querySelector('img');
