+++ static/MelaoApp/js/newPost.js
@@ -1,14 +1,12 @@
 function setTheme(){
     const body = document.body;
+    const Xicon = document.getElementById('X-Icon');
 
     const applyTheme = (theme) => {
         if (theme === 'dark') {
-            body.classList.add('dark-mode');
-            localStorage.setItem('theme', 'dark');
-            
+            Xicon.querySelector('img').src = '/MelaoApp/resources/DarkMode/X-Icon.png';
         } else {
-            body.classList.remove('dark-mode');
-            localStorage.setItem('theme', 'light');
+            Xicon.querySelector('img').src = '/MelaoApp/resources/X-Icon.png';
         }
     };
